refactored CloudSdkPreconditionsSupport to CloudSdkServiceManager.
