+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/appengine/sdk/CloudSdkServiceManagerTest.java
@@ -1,19 +1,3 @@
-/*
- * Copyright 2018 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.google.cloud.tools.intellij.appengine.sdk;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -39,8 +23,8 @@ import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.Spy;
 
-/** Tests for {@link CloudSdkPreconditionsSupport}. */
-public class CloudSdkPreconditionsSupportTest {
+/** Tests for {@link CloudSdkServiceManager}. */
+public class CloudSdkServiceManagerTest {
   @Rule public CloudToolsRule cloudToolsRule = new CloudToolsRule(this);
 
   @Mock @TestService private CloudSdkServiceManager mockCloudSdkServiceManager;
@@ -50,7 +34,7 @@ public class CloudSdkPreconditionsSupportTest {
   @Mock private Project mockProject;
   @Mock private CloudSdkPreconditionCheckCallback mockCallback;
 
-  @Spy private CloudSdkPreconditionsSupport cloudSdkPreconditionsSupport;
+  @Spy private CloudSdkServiceManager cloudSdkServiceManager;
 
   @Before
   public void setUp() {
@@ -61,7 +45,7 @@ public class CloudSdkPreconditionsSupportTest {
   public void installingSdk_then_readySdk_correctly_runs() {
     mockSdkStatusChange(SdkStatus.INSTALLING, SdkStatus.READY);
 
-    cloudSdkPreconditionsSupport.runAfterCloudSdkPreconditionsMet(
+    cloudSdkServiceManager.runAfterCloudSdkPreconditionsMet(
         mockProject, mockRunnable, "", mockCallback);
 
     ApplicationManager.getApplication().invokeAndWait(() -> verify(mockRunnable).run());
@@ -71,9 +55,9 @@ public class CloudSdkPreconditionsSupportTest {
   public void installingSdk_then_stillInstalling_doesNotRun() {
     mockSdkStatusChange(SdkStatus.INSTALLING, SdkStatus.INSTALLING);
     // mock cancel operation for incomplete install.
-    doReturn(true).when(cloudSdkPreconditionsSupport).checkIfCancelled();
+    doReturn(true).when(cloudSdkServiceManager).checkIfCancelled();
 
-    cloudSdkPreconditionsSupport.runAfterCloudSdkPreconditionsMet(
+    cloudSdkServiceManager.runAfterCloudSdkPreconditionsMet(
         mockProject, mockRunnable, "", mockCallback);
 
     ApplicationManager.getApplication().invokeAndWait(() -> verifyNoMoreInteractions(mockRunnable));
@@ -83,7 +67,7 @@ public class CloudSdkPreconditionsSupportTest {
   public void installingSdk_then_invalidSdk_doesNotRun() {
     mockSdkStatusChange(SdkStatus.INSTALLING, SdkStatus.INVALID);
 
-    cloudSdkPreconditionsSupport.runAfterCloudSdkPreconditionsMet(
+    cloudSdkServiceManager.runAfterCloudSdkPreconditionsMet(
         mockProject, mockRunnable, "", mockCallback);
 
     ApplicationManager.getApplication().invokeAndWait(() -> verifyNoMoreInteractions(mockRunnable));
@@ -93,7 +77,7 @@ public class CloudSdkPreconditionsSupportTest {
   public void installingSdk_then_notAvailableSdk_doesNotRun() {
     mockSdkStatusChange(SdkStatus.INSTALLING, SdkStatus.NOT_AVAILABLE);
 
-    cloudSdkPreconditionsSupport.runAfterCloudSdkPreconditionsMet(
+    cloudSdkServiceManager.runAfterCloudSdkPreconditionsMet(
         mockProject, mockRunnable, "", mockCallback);
 
     ApplicationManager.getApplication().invokeAndWait(() -> verifyNoMoreInteractions(mockRunnable));
@@ -103,16 +87,16 @@ public class CloudSdkPreconditionsSupportTest {
   public void installingSdk_then_userCancel_doesNotShowWarningNotification() {
     mockSdkStatusChange(SdkStatus.INSTALLING, SdkStatus.INSTALLING);
     // mock cancel operation for incomplete install.
-    doReturn(true).when(cloudSdkPreconditionsSupport).checkIfCancelled();
+    doReturn(true).when(cloudSdkServiceManager).checkIfCancelled();
 
-    cloudSdkPreconditionsSupport.runAfterCloudSdkPreconditionsMet(
+    cloudSdkServiceManager.runAfterCloudSdkPreconditionsMet(
         mockProject, mockRunnable, "", mockCallback);
 
     ApplicationManager.getApplication()
         .invokeAndWait(
             () ->
                 // explicit parameters are not relevant but need to be passed for spy to work.
-                verify(cloudSdkPreconditionsSupport, never())
+                verify(cloudSdkServiceManager, never())
                     .showCloudSdkNotification("", NotificationType.WARNING, false));
   }
 
@@ -120,13 +104,13 @@ public class CloudSdkPreconditionsSupportTest {
   public void installingSdk_then_invalidSdk_showsErrorNotification() {
     mockSdkStatusChange(SdkStatus.INSTALLING, SdkStatus.INVALID);
 
-    cloudSdkPreconditionsSupport.runAfterCloudSdkPreconditionsMet(
+    cloudSdkServiceManager.runAfterCloudSdkPreconditionsMet(
         mockProject, mockRunnable, "", mockCallback);
 
     ApplicationManager.getApplication()
         .invokeAndWait(
             () ->
-                verify(cloudSdkPreconditionsSupport)
+                verify(cloudSdkServiceManager)
                     .showCloudSdkNotification(
                         GctBundle.message("appengine.deployment.error.sdk.invalid"),
                         NotificationType.ERROR,
